Disallow `null` as configuration attribute value
The rationale for this change is that we now have the `org.gradle.api.artifacts.ConfigurationAttributeMatcher#defaultValue`
method which is responsible for returning a default value in case an attribute is missing. The return type of this method
is `String`, which makes it impossible to determine if the default value is `null`, or if the matcher doesn't provide a
default value. The algorithm assumes that `null` as the result of `defaultValue` means that the matcher doesn't provide
a default value. Therefore, we would have a problem in case we want to match on the `null` value.

To fix this, we decided not to introduce the `Optional` API, because even if it would solve the problem, we currently
don't have access to it: either we would need to use the Guava type, which would mean exposing a Guava type to a public
API (we don't want this), or using Java 8's `Optional`, but Gradle still needs to run on older JDKs.

In order to avoid a more complex API, we therefore decided to prevent using `null` as an attribute value.

See: gradle/performance#86
+review REVIEW-6358